chore: Add todo list and todo list item components
This commit adds the todo list and todo list item components to the application. The todo list component is responsible for displaying a list of todo items, while the todo list item component represents an individual todo item.

The components are implemented in both the template files (todo-list.component.html and todo-list-item.component.html) and the style files (todo-list.component.less and todo-list-item.component.less).

The todo list component is imported and used in the app component (app.component.ts) to display the list of todo items in the application.

This commit also includes test files (todo-list.component.spec.ts and todo-list-item.component.spec.ts) for both components, which contain basic tests to ensure that the components are created successfully.