feat: Add add-todo component
This commit adds a new add-todo component to the application. The add-todo component is responsible for providing a form to add new todo items. It includes an input field to enter the todo text and a submit button to add the todo item.

The add-todo component is implemented in the template file (add-todo.component.html) and the style file (add-todo.component.less). It also includes a TypeScript file (add-todo.component.ts) that handles the logic for capturing user input and emitting an event when a new todo item is added.

This commit also includes test files (add-todo.component.spec.ts) for the add-todo component, which contain basic tests to ensure that the component is created successfully and that the add-todo functionality works as expected.